# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ShortMesh Interface service",
        "title": "Shortmesh - Interface API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/api-keys": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all API keys for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikeys"
                ],
                "summary": "List API keys",
                "responses": {
                    "200": {
                        "description": "List of API keys",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apikeys.APIKeyInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apikeys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apikeys.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new API key for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikeys"
                ],
                "summary": "Create a new API key",
                "parameters": [
                    {
                        "description": "API key creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apikeys.CreateAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "API key created successfully",
                        "schema": {
                            "$ref": "#/definitions/apikeys.APIKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/apikeys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apikeys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apikeys.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an API key for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apikeys"
                ],
                "summary": "Delete an API key",
                "parameters": [
                    {
                        "description": "API key ID to delete",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apikeys.DeleteAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "API key deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/apikeys.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/apikeys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apikeys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "API key not found",
                        "schema": {
                            "$ref": "#/definitions/apikeys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apikeys.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all devices for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "List all devices",
                "responses": {
                    "200": {
                        "description": "List of devices",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/devices.Device"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new device for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Create a new device",
                "parameters": [
                    {
                        "description": "Device creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devices.CreateDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Requested to add device successfully",
                        "schema": {
                            "$ref": "#/definitions/devices.DeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a device for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Delete a device",
                "parameters": [
                    {
                        "description": "Device deletion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devices.DeleteDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/devices.DeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/devices/qr-code": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Establishes a WebSocket connection to stream real-time add devices qr-code. This endpoint cannot be tested in Swagger UI - use a WebSocket client instead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "WebSocket qr-code endpoint (Not executable in Swagger UI)",
                "deprecated": true,
                "responses": {
                    "101": {
                        "description": "WebSocket connection established",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid authentication token",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apikeys.APIKeyInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-19T20:00:00Z"
                },
                "expires_at": {
                    "type": "string",
                    "example": "2026-12-31T23:59:59Z"
                },
                "key_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6g7h8"
                },
                "last_used_at": {
                    "type": "string",
                    "example": "2026-02-19T20:30:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Production API Key"
                }
            }
        },
        "apikeys.APIKeyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/apikeys.APIKeyInfo"
                },
                "key": {
                    "type": "string",
                    "example": "ak_123456789abcdef123456789abcdef123456789"
                },
                "message": {
                    "type": "string",
                    "example": "API key created successfully"
                }
            }
        },
        "apikeys.CreateAPIKeyRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2026-12-31T23:59:59Z"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Production API Key"
                }
            }
        },
        "apikeys.DeleteAPIKeyRequest": {
            "type": "object",
            "required": [
                "key_id"
            ],
            "properties": {
                "key_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6g7h8"
                }
            }
        },
        "apikeys.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "apikeys.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "API key deleted successfully"
                }
            }
        },
        "devices.CreateDeviceRequest": {
            "type": "object",
            "required": [
                "platform"
            ],
            "properties": {
                "platform": {
                    "type": "string",
                    "example": "wa"
                }
            }
        },
        "devices.DeleteDeviceRequest": {
            "type": "object",
            "required": [
                "device_id",
                "platform"
            ],
            "properties": {
                "device_id": {
                    "description": "Get the device ID from ListDevices (GET /api/v1/devices) response",
                    "type": "string",
                    "example": "237123456789"
                },
                "platform": {
                    "description": "Get the platform from ListDevices (GET /api/v1/devices) response",
                    "type": "string",
                    "example": "wa"
                }
            }
        },
        "devices.Device": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string",
                    "example": "237123456789"
                },
                "platform": {
                    "type": "string",
                    "example": "wa"
                }
            }
        },
        "devices.DeviceResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Scan the QR code to link your device"
                },
                "qr_code_url": {
                    "type": "string",
                    "example": "/api/v1/devices/qr-code"
                }
            }
        },
        "devices.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}